bugfix: avoid to recalculate digest for existing layer
for the legacy schema1 image, the digest info about layer is not equal
to the existing blob data. fetchBlob needs to calculate the digest for
the layer blob data. however, containerd doesn't cache the result. for
the existing blob data, fetchBlob function always calculate the digest
when re-pull the same image.

based on this case, we can use label to cache the result to avoid to
recalculate digest when re-pull the same image.